[Gems.Http] Исправил логирование длительности http запроса

+++ src/Logging/Mvc/src/LogsCollector/RequestLogsCollector.cs
@@ -113,7 +113,7 @@ namespace Gems.Logging.Mvc.LogsCollector
             this.AddOrUpdateValueInLogs(key, value);
         }
 
-        public void AddRequestDuration(long durationMilliseconds)
+        public void AddRequestDuration(double durationMilliseconds)
         {
             this.AddOrUpdateValueInLogs("requestDuration", durationMilliseconds);
         }
